Fix internal dependencies
`wxc` and `wxdirect` are not Haskell libraries, so they are no longer registered as such.

The `--wxc` flag to `wxdirect` is now used to point to the location of the header files. Unfortunately this is hardcoded for now in `wxcore`'s `Setup.hs`.

The intention is that `wxc` is installed as a system-level dependency, but we haven't determined how best to do this yet. For now, it is required to manually shuffle around some files:
- `cabal build wxc`
- `find . -name libwxc.so.0.93.0.0 -exec sudo cp {} /usr/lib/libwxc.so \;`
- `sudo cp -r wxc/src/include /usr/include/wxc`