wx: Fix event handling for wxWidgets > 3.0
In the example `cabal run samplesWx:BouncingBalls`, we had

```haskell
       set p [on click         := dropBall vballs p             -- drop ball
             ,on clickRight    := (\_pt -> ballsFrame)          -- new window
             ,on (charKey 'p') := set t [enabled :~ not]        -- pause
             ,on (charKey '-') := set t [interval :~ \i -> i*2] -- increase interval
             ,on (charKey '+') := set t [interval :~ \i -> max 1 (i `div` 2)]  -- decrease interval
             ]
```

With the above code only the latest even handler for an event was
being registered. So only the right click or `+` key would work.

The haskell code registers event handlers on the window for a given
event and when a handler already exists it should fetch the available
handler and add to it. However, the function `evtHandlerGetClosure`
was returning `null` even when there was an existing event. I found
out that this was because the function `wxEvtHandler_GetClosure` in
`wrapper.cpp` was disabled on wxWidgets >= 3.1 as `wxEvtHandler`
class no longer had a method called `GetDynamicEventTable`.

I had to use a new internal function `GetFirstDynamicEntry` to check
for that dynamic events are enabled and then search for our event.
Without the check, `SearchDynamicEventTable` throws a debug window.

Relevant wxWidgets commit that introduced the change:
https://github.com/wxWidgets/wxWidgets/commit/99d9a81e28e93106b2a471198cbfaa474a0fa714